python_homework/requirements.txt
Flask
SQLAlchemy
python-json-logger
gunicorn
prometheus_client
redis
PyJWT
flake8
kafka-python